独立开发删除异常报错接口

+++ Tlias-Test/src/main/java/org/example/tliastest/service/impl/ClazzServiceImpl.java
@@ -2,13 +2,18 @@ package org.example.tliastest.service.impl;
 
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
+import org.example.tliastest.exception.DelClazzException;
 import org.example.tliastest.mapper.ClazzMapper;
+import org.example.tliastest.mapper.StudentsMapper;
 import org.example.tliastest.pojo.Clazz;
 import org.example.tliastest.pojo.ClazzQueryParam;
+import org.example.tliastest.pojo.EmpLog;
 import org.example.tliastest.pojo.PageResult;
 import org.example.tliastest.service.ClazzService;
+import org.example.tliastest.service.EmpLogService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -19,6 +24,12 @@ public class ClazzServiceImpl implements ClazzService {
 
     @Autowired
     ClazzMapper clazzMapper;
+
+    @Autowired
+    StudentsMapper studentsMapper;
+
+    @Autowired
+    EmpLogService empLogService;
 
     @Override
     public PageResult<Clazz> getClassList(ClazzQueryParam clazzQueryParam) {
@@ -40,9 +51,19 @@ public class ClazzServiceImpl implements ClazzService {
         return pageResult;
     }
 
+    @Transactional(rollbackFor = Exception.class)
     @Override
-    public void delClass(Integer id) {
-        clazzMapper.deleteByPrimaryKey(id);
+    public void delClass(Integer id){
+        try {
+            Integer count = studentsMapper.selectByClazzId(id);
+            if (count > 0) {
+                throw new DelClazzException("id为" + id + "的班级下有" + count + "个学生，不能删除");
+            }
+            clazzMapper.deleteByPrimaryKey(id);
+        } finally {
+            EmpLog log = new EmpLog(null, LocalDateTime.now(), "删除班级" + id);
+            empLogService.insertLog(log);
+        }
     }
 
     @Override
